# Conflicts:
#	app/src/main/java/ir/fatemelyasi/note/model/local/dao/LabelDao.kt
#	app/src/main/java/ir/fatemelyasi/note/model/local/mappers/NoteMapper.kt
#	app/src/main/java/ir/fatemelyasi/note/model/noteLocalDataSource/NoteLocalDataSource.kt
#	app/src/main/java/ir/fatemelyasi/note/model/noteLocalDataSource/NoteLocalDataSourceImpl.kt
#	app/src/main/java/ir/fatemelyasi/note/model/noteLocalRepository/NoteLocalRepository.kt
#	app/src/main/java/ir/fatemelyasi/note/model/noteLocalRepository/NoteLocalRepositoryImpl.kt

+++ app/src/main/java/ir/fatemelyasi/note/model/local/entity/CrossEntity.kt
@@ -26,7 +26,7 @@ import androidx.room.Index
 )
 data class CrossEntity(
     @ColumnInfo("noteId")
-    val noteId: Int,
+    val noteId: Long,
     @ColumnInfo("labelId")
-    val labelId: Int
+    val labelId: Long
 )
